integration test folder restructure move wait-for commands to script add extra package.json scripts
