+++ package.json
@@ -10,6 +10,8 @@
     "test": "jest --runInBand",
     "test:lint": "standard",
     "test:pact": "jest --testPathPattern=test/contract --coverage=false",
+    "test:integration": "jest --testPathPattern=test/integration --coverage=false",
+    "test:unit": "jest --testPathPattern=test/unit --coverage=false",
     "test:watch": "jest --coverage=false --onlyChanged --watch --runInBand",
     "start:watch": "nodemon --inspect=0.0.0.0 --ext js --legacy-watch index.js"
   },
