+++ scripts/wait-for-dependencies
@@ -1,0 +1,3 @@
+scripts/wait-for/wait-for ffc-demo-payment-postgres:5432 --
+scripts/wait-for/wait-for ffc-demo-payment-message-queue:9324 --
+npm run migrate
