+++ server/plugins/auth/verify-audience.js
@@ -1,6 +1,6 @@
 function verifyAudience (audience, expectedAudience) {
   if (audience !== expectedAudience) {
-    throw Error(`Invalid audience: ${audience}`)
+    throw Error(`Invalid audience: ${audience} is not ${expectedAudience}`)
   }
 }
 module.exports = verifyAudience
