+++ test/integration/local/services/payment-service.test.js
@@ -1,5 +1,5 @@
-const dbHelper = require('../../db-helper')
-const paymentService = require('../../../server/services/payment-service')
+const dbHelper = require('../../../db-helper')
+const paymentService = require('../../../../server/services/payment-service')
 
 const payment1 = {
   claimId: 'MINE123',
