+++ test/integration/narrow/routes/healthz.test.js
@@ -3,8 +3,8 @@ describe('Healthz test', () => {
   let server
 
   beforeAll(async () => {
-    jest.mock('../../../server/services/message-service')
-    createServer = require('../../../server')
+    jest.mock('../../../../server/services/message-service')
+    createServer = require('../../../../server')
   })
 
   beforeEach(async () => {
