+++ test/integration/narrow/routes/payment.test.js
@@ -4,10 +4,10 @@ describe('Payment test', () => {
   let paymentService
 
   beforeAll(async () => {
-    jest.mock('../../../server/services/message-service')
-    jest.mock('../../../server/services/payment-service')
-    paymentService = require('../../../server/services/payment-service')
-    createServer = require('../../../server')
+    jest.mock('../../../../server/services/message-service')
+    jest.mock('../../../../server/services/payment-service')
+    paymentService = require('../../../../server/services/payment-service')
+    createServer = require('../../../../server')
   })
 
   beforeEach(async () => {
