+++ test/integration/narrow/server/index.test.js
@@ -3,32 +3,32 @@ describe('Server tests', () => {
   let server
 
   beforeAll(() => {
-    jest.mock('../../../server/services/message-service')
+    jest.mock('../../../../server/services/message-service')
   })
 
   test('createServer returns server', () => {
-    jest.mock('../../../server/config', () => {
+    jest.mock('../../../../server/config', () => {
       return {
         port: 3004,
         env: 'production'
       }
     })
 
-    createServer = require('../../../server')
+    createServer = require('../../../../server')
     server = createServer()
 
     expect(server).toBeDefined()
   })
 
   test('createServer returns server in development', () => {
-    jest.mock('../../../server/config', () => {
+    jest.mock('../../../../server/config', () => {
       return {
         port: 3004,
         isDev: true
       }
     })
 
-    createServer = require('../../../server')
+    createServer = require('../../../../server')
     server = createServer()
 
     expect(server).toBeDefined()
